Student_management/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
postgres==4.0
psycopg==3.0.8
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2